Fixing disk issue which was causing the sdk to wait for the wrong condition.
When a new disk was added and other disks where removed at the same
time, the sdk was considering the removed disks as new disks. The
consequence of this was that the payload comparision (
UpdateVMRequest.compareResponse() ) would never converge to desired
state making the sdk retry until timeout is reached.